add/remove fighters from pools
