+++ app/models/fighter.rb
@@ -13,10 +13,6 @@
 
 class Fighter < ActiveRecord::Base
   validates :first_name, :last_name, presence: true
-  has_many :tournament_fighters
-  has_many :tournaments,
-    through: :tournament_fighters,
-    source: :tournament
   has_many :match_fighters,
     class_name: 'MatchFighter',
     primary_key: :id,
