+++ app/models/tournament.rb
@@ -18,10 +18,6 @@ class Tournament < ActiveRecord::Base
   validates :event_id, :name, :weapon_type, presence: true
   belongs_to :event
   has_many :scores, dependent: :destroy
-  has_many :tournament_fighters
-  has_many :fighters,
-    through: :tournament_fighters,
-    source: :fighter
   has_many :pools, dependent: :destroy
 
   after_create :create_default_pool, :add_score
@@ -32,22 +28,5 @@ class Tournament < ActiveRecord::Base
 
   def add_score
     scores.create(target: 'head', points: 4)
-  end
-
-  def add_or_create_fighter(first_name, last_name)
-    fighter = Fighter.where(first_name: first_name, last_name: last_name)
-
-    if fighter.empty?
-      fighters.create(first_name: first_name, last_name: last_name)
-      "#{first_name} #{last_name} created and added"
-    else
-      tournament_fighters.create(fighter_id: fighter.first.id)
-      "#{first_name} #{last_name} added"
-    end
-  end
-
-  def remove_fighter(fighter_id)
-    tournament_fighters.find_by_fighter_id(fighter_id).destroy
-    pools.each { |pool| pool.remove_fighter(fighter_id) }
   end
 end
