+++ db/migrate/20150818040910_remove_match_fighters.rb
@@ -1,0 +1,5 @@
+class RemoveMatchFighters < ActiveRecord::Migration
+  def change
+    drop_table :tournament_fighters
+  end
+end
